Add files via upload
The project is called "ECommerce Sales Analysis" and it presents a comprehensive analysis of an e-commerce business, utilizing SQL and Power BI tools. The project analyzes key performance metrics and trends. The insights gained from this analysis can aid in informed decision-making and drive business growth.
The aim of this project is to provide a clear and actionable understanding of the e-commerce business and its various components, helping to identify areas for improvement and growth.